// app/types/completeprojects.ts
// export interface completeprojects {
//     id: string;
//     title: string;
//     description: string;
//     status: "New" | "In Progress" | "Completed";
//     imageUrl: string;
//     duration: string;
//     completedDate: string;
// }
//
export interface completeprojects {
    id: string;
    title: string;
    description: string;
    status: "New" | "In Progress" | "Completed";
    imageUrl: string;
    duration: string;
    completedDate: string;
    visibility: "Visible" | "Hidden";
}